feat: move deal-detail page to offer page and remove demand, swap detail
