--- app/(deal-detail)/demand/[slug]/page.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import DemandDetail from '@/components/demands/demand-detail';
-
-function DemandPage({
-  params: { slug },
-}: {
-  params: {
-    slug: string;
-  };
-}) {
-  return <DemandDetail slug={slug} />;
-}
-
-export default DemandPage;
