--- app/(deal-detail)/offer/[slug]/page.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import OfferDetail from '@/components/offers/offer-detail';
-
-function OfferPage({
-  params: { slug },
-}: {
-  params: {
-    slug: string;
-  };
-}) {
-  return <OfferDetail slug={slug} />;
-}
-
-export default OfferPage;
