--- app/(deal-detail)/swap/[slug]/page.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import SwapDetail from '@/components/swaps/swap-detail';
-
-function SwapPage({
-  params: { slug },
-}: {
-  params: {
-    slug: string;
-  };
-}) {
-  return <SwapDetail slug={slug} />;
-}
-
-export default SwapPage;
