+++ app/offer/[slug]/report/layout.tsx
