+++ app/offer/[slug]/report/page.tsx
